moving LibraryStats struct into shared module
